+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - jslint - try to improve parser to be able to parse jquery.js without stopping.
 
 # v2026.3.1-beta
+- jslint-ci - Update file jslint_ci.sh, replacing deprecated url.parse(...).pathname with new URL(...).pathname.
 - jslint-ci - Update shell-function shGitPullrequestCleanup() to allow squash-and-merge pull-request.
 - ci - Update file .github/workflows/ci.yml with hook to run on pull-request.
 - jslint-ci - Update shell-function shLintPython().